Changed energy_gap threshold in invest_energy_technologies.py
Matlab and Python diverge in very small numbers for energy_gap, this leads to different results in coord_gap due to >0. Changed this to threshold of 1e-12 to exclude these very small numbers on both sides. Note: Maybe this needs to be changed back later, I need to check if this is really justifiable.